feat: 친구 목록 css 설정 (#5)
