+++ src/styles/FriendItem.css
@@ -1,0 +1,35 @@
+.freind-item-container{
+  width: 290px;
+  margin-left: 5px;
+  margin-top: 10px;
+  height: 43px;
+  border-radius: 10px;
+  border: 0.7px solid gray;
+  box-shadow: 0 2px 2px rgba(0,0,0,0.25), 0 2px 2px rgba(0,0,0,0.22);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+.freind-item-sub-container{
+  width: 270px;
+  display: flex;
+}
+.friend-item-on-off-line{
+  width: 20px;
+  height: 20px;
+  background-color: red;
+  border-radius: 10px;
+}
+.friend-item-nickname-container{
+  width: 250px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+.friend-item-nickname{
+  font-size: 20px;
+}
+
+.freind-item-sub-container:active{
+  opacity: 0.7;
+}
