+++ src/styles/FriendList.css
@@ -1,18 +1,21 @@
 .freind-list-container{
-  width: 100%;
-  height: 100%;
+  width: 300px;
+  height: 700px;
+  border-radius: 30px;
+  box-shadow: 0 10px 10px rgba(0,0,0,0.25), 0 10px 10px rgba(0,0,0,0.22);
 }
 .freind-list-search-button{
   width: 100%;
   height: 43px;
   background-color: #8ABEFF;
   border: 0.7px solid gray;
+  border-radius: 30px 30px 0 0;
 }
 .freind-list-search-button:active{
   opacity: 0.7;
 }
 .freind-list-item-container{
   width: 312px;
-  height: 689px;
+  height: 650px;
   overflow: scroll;
 }
