+++ src/styles/Sidebar.css
@@ -40,6 +40,9 @@
   width: 312px;
   height: 731px;
   background-color:rgb(255, 255, 255);
+  display: flex;
+  justify-content: center;
+  align-items: center;
 }
 .sidebar-setting-button{
   width: 312px;
